Add template code to start populating the cards with printers.
When adding a printer, it asks for a deparment name which will be used
as a heading in the site to put all the printers in that department
together.
Then, it creates a card for every printer already in the database under
that department.
If no printers are in the database, then a message is displayed saying
to click on the add printer button to start using the program