Change way to get toner name and add exceptions.
Previously, the toner name was the actual full name with the toner
model (Black Cartridge 48A (CF248A)). Now the toner name is just the
color.

Easysnmp throws another exception (EasySNMPTimeoutError) when trying to
use a different version of snmp. Sometimes I get SystemError and other
times EasySNMPTimeoutError, not sure why but added it just to make sure
it works.